projects/blue team/requirements.txt
# Web Framework and Server
fastapi==0.136.1
uvicorn[standard]==0.46.0
python-multipart==0.0.27
websockets==16.0

# Web Scraping and Networking
beautifulsoup4==4.14.3
requests==2.33.1
PySocks==1.7.1

# Environment and Configuration
python-dotenv==1.2.2

# AI and LLM Integrations
langchain-core==1.3.2
langchain-community==0.4.1
langchain-openai==1.2.1
langchain-ollama==1.1.0
langchain-anthropic==1.4.2
langchain-google-genai==4.2.2
openai==2.33.0